Fixed compilation issues on Windows 64-bit: set /bigobj for all builds of cisstMultiTask; modified cisstvsvars.bat to call vcvarsall.bat (if it exists) instead of vsvars32.bat; removed erroneous dependencies on cisstNumerical; removed some warnings.
git-svn-id: https://svn.lcsr.jhu.edu/cisst/trunk@3822 0959eb1a-6956-4f60-947b-6092a4e22c97

+++ cisst/cisstOSAbstraction/code/osaPipeExec.cpp
@@ -454,7 +454,7 @@ int osaPipeExec::Write(const char * buffer, int n)
 
 int osaPipeExec::Write(const char * buffer)
 {
-    return Write(buffer, strlen(buffer)+1);
+    return Write(buffer, static_cast<int>(strlen(buffer))+1);
 }
